article-service/pyproject.toml
[project]
name = "article-service"
version = "1.0.0"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "redis==5.0.1",
    "psycopg2-binary==2.9.9",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
]

[build-system]
requires = ["setuptools", "wheel"]

[tool.setuptools.packages.find]
where = ["."]  